Allowed for multiple client-side scripts
- Renamed ArticleContainer to Page
- Migrated some functionality from server to Page
