+++ src/client/liveblog.ts
@@ -1,0 +1,8 @@
+// ----- Imports ----- //
+
+import setup from 'client/setup';
+
+
+// ----- Run ----- //
+
+setup();
